+++ frontend/api/api.js
@@ -1,4 +1,5 @@
 import axios from 'axios'
+import { Base64 } from 'js-base64'
 
 const api = {
   getConfig() {
@@ -187,6 +188,23 @@ const api = {
         newpassword: params.newpassword,
       })
         .then(res => resolve(res.data.data))
+        .catch(error => reject(error))
+    })
+  },
+  downloadItem (params) {
+    return new Promise((resolve, reject) => {
+      axios.get('download&path='+encodeURIComponent(Base64.encode(params.path)))
+        .then(res => resolve(res.data))
+        .catch(error => reject(error))
+    })
+  },
+  saveContent (params) {
+    return new Promise((resolve, reject) => {
+      axios.post('savecontent', {
+        name: params.name,
+        content: params.content,
+      })
+        .then(res => resolve(res.data))
         .catch(error => reject(error))
     })
   },