+++ frontend/translations/german.js
@@ -70,6 +70,10 @@ const data = {
   'Updated': 'Aktualisiert',
   'Deleted': 'Gelöscht',
   'Your file is ready': 'Deine Datei ist fertig',
+  'View': 'View',
+  'Search': 'Search',
+  'Download permission': 'Herunterladen',
+  'Guest': 'Guest',
 }
 
 export default data