+++ frontend/translations/turkish.js
@@ -70,6 +70,10 @@ const data = {
   'Updated': 'Güncellendi',
   'Deleted': 'Silindi',
   'Your file is ready': 'Dosyanız Hazır',
+  'View': 'View',
+  'Search': 'Search',
+  'Download permission': 'İndir',
+  'Guest': 'Guest',
 }
 
 export default data