+++ frontend/views/Login.vue
@@ -13,10 +13,10 @@
             </div>
             <br>
             <b-field :label="lang('Username')">
-              <b-input v-model="username" name="username" required @input="error = ''" />
+              <b-input v-model="username" name="username" required @input="error = ''" ref="username" />
             </b-field>
             <b-field :label="lang('Password')">
-              <b-input v-model="password" type="password" name="password" required @input="error = ''" />
+              <b-input v-model="password" type="password" name="password" required @input="error = ''" password-reveal />
             </b-field>
 
             <div class="is-flex is-justify-end">
@@ -46,6 +46,9 @@ export default {
       password: '',
       error: '',
     }
+  },
+  mounted() {
+    this.$refs.username.focus()
   },
   methods: {
     login() {